CHROMIUM: hrtimer: Disable high resolution timers
Finch experiment was going on for the last few months with highres
timers disabled and results are generally positive:

124 beta 100%:
https://uma.googleplex.com/p/chrome/variations?sid=a16be03f03c14b440edd235ef6554d93

123 stable 1%:
https://uma.googleplex.com/p/chrome/variations?sid=11f97e0da078a4af95ae7b465c913ee3

Beta has more data as experiment is enabled 100%.

Disable high resolution timers permanently on 5.15. This will be
followed by moving finch experiment to 100% stable.

This is done only for x86 as of now. More verification is needed on arm boards before we rollout this feature on non-x86 boards.

BUG=b:263289152
TEST=boot
Cq-Depend: chromium:5730074
Disallow-Recycled-Builds: all
Change-Id: I411f4b5792c163466e218f960eed8e228267d7cd
Signed-off-by: Vineeth Pillai <vineethrp@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5570389
Reviewed-by: Joel Fernandes <joelaf@google.com>