UPSTREAM: eventfs: Do not use attributes for events directory
The top "events" directory has a static inode (it's created when it is and
removed when the directory is removed). There's no need to use the events
ei->attr to determine its permissions. But it is used for saving the
permissions of the "events" directory for when it is created, as that is
needed for the default permissions for the files and directories
underneath it.

For example:

 # cd /sys/kernel/tracing
 # mkdir instances/foo
 # chown 1001 instances/foo/events

The files under instances/foo/events should still have the same owner as
instances/foo (which the instances/foo/events ei->attr will hold), but the
events directory now has owner 1001.

Link: https://lore.kernel.org/lkml/20240522165032.104981011@goodmis.org

BUG=b:360162015
TEST=find /sys/kernel/tracing

Cc: Linus Torvalds <torvalds@linux-foundation.org>
Cc: Masami Hiramatsu <mhiramat@kernel.org>
Cc: Mark Rutland <mark.rutland@arm.com>
Cc: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Cc: Andrew Morton <akpm@linux-foundation.org>
Change-Id: Ib056c771cf12a16dc549cdafe6daa242f219d6e8
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>
(cherry picked from commit 2dd00ac1d38afba1b59e439abc300a9b0ce696bf)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5792771
Reviewed-by: Sarthak Kukreti <sarthakkukreti@google.com>
Reviewed-by: Sean Paul <sean@poorly.run>
Auto-Submit: Steven Rostedt <rostedt@google.com>
Tested-by: Steven Rostedt <rostedt@google.com>
Commit-Queue: Steven Rostedt <rostedt@google.com>