UPSTREAM: eventfs: Update all the eventfs_inodes from the events descriptor
The change to update the permissions of the eventfs_inode had the
misconception that using the tracefs_inode would find all the
eventfs_inodes that have been updated and reset them on remount.
The problem with this approach is that the eventfs_inodes are freed when
they are no longer used (basically the reason the eventfs system exists).
When they are freed, the updated eventfs_inodes are not reset on a remount
because their tracefs_inodes have been freed.

Instead, since the events directory eventfs_inode always has a
tracefs_inode pointing to it (it is not freed when finished), and the
events directory has a link to all its children, have the
eventfs_remount() function only operate on the events eventfs_inode and
have it descend into its children updating their uid and gids.

Link: https://lore.kernel.org/all/CAK7LNARXgaWw3kH9JgrnH4vK6fr8LDkNKf3wq8NhMWJrVwJyVQ@mail.gmail.com/
Link: https://lore.kernel.org/linux-trace-kernel/20240523051539.754424703@goodmis.org

BUG=b:360162015
TEST=find /sys/kernel/tracing

Cc: stable@vger.kernel.org
Cc: Masami Hiramatsu <mhiramat@kernel.org>
Cc: Mark Rutland <mark.rutland@arm.com>
Cc: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Cc: Andrew Morton <akpm@linux-foundation.org>
Fixes: baa23a8d4360d ("tracefs: Reset permissions on remount if permissions are options")
Reported-by: Masahiro Yamada <masahiroy@kernel.org>
Change-Id: Ie95d81de48c5d6dcafa886afece0ece25c5c3ec0
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>
(cherry picked from commit 340f0c7067a95281ad13734f8225f49c6cf52067)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5792766
Tested-by: Steven Rostedt <rostedt@google.com>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@google.com>
Auto-Submit: Steven Rostedt <rostedt@google.com>
Commit-Queue: Steven Rostedt <rostedt@google.com>
Reviewed-by: Sean Paul <sean@poorly.run>