CHROMIUM: PMR physical memory is freed before GPU TLB invalidation
Instead of freeing the pages, the factory implementation instead return
the list of pages to pmr.c, where they will be added to the a list to be
freed later.

This changes also adds a check to PhysMemValidateParams() to reject any
request to create a sparse allocation which is backed ON_DEMAND as this
combination is invalid.

IMG Ref: PP159595
BUG= b/339665658
TEST=graphics.DRM.*, graphics.WebGLAquarium.1000_fishes,
     graphics.WebGLManyPlanetsDeep, graphics.GLBench

Cq-Depend: chromium:5666794
Change-Id: Ia0f6662648e7f95511c51fdbc6f68db8bed60273
Signed-off-by: Faijur Rahman <faijur.rahman@imgtec.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5666795
Reviewed-by: Pin-yen Lin <treapking@chromium.org>
Tested-by: Pin-yen Lin <treapking@chromium.org>
Commit-Queue: Pin-yen Lin <treapking@chromium.org>