CHROMIUM: wifi: mac80211: Do not return released skb from ieee80211_tx_dequeue
ieee80211_tx_dequeue() may release an skb but return it anyway
if the transmit queue is stopped during the call.

Prevent this from happening by clearing the skb whenever restarting
the function.

CHROMIUM note:

This problem has been fixed differently in the upstream kernel.
Backporting the fixes would be difficult. This is a downstream only
fix which is not needed in the upstream kernel.

BUG=b:348944088, b:347452142
TEST=CQ, wireless tests

Change-Id: If49ba0f447f3a5fd792cf22ae61338d419a389cf
Disallow-Recycled-Builds: test-failures
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5660958
Commit-Queue: Martin Faltesek <mfaltesek@google.com>
Reviewed-by: Martin Faltesek <mfaltesek@google.com>