CHROMIUM: KVM: x86: Include host suspended time in steal time.
When the host resumes from a suspend, the guest thinks any task
that was running during the suspend ran for a long time, even though
the effective run time was much shorter, which can end up having
negative effects with scheduling. This can be particularly noticeable
if the guest task was RT, as it can end up getting throttled for a
long time.

To mitigate this issue, we include the time that the host was
suspended in steal time, which lets the guest can subtract the
duration from the tasks' runtime.

(based on https://lore.kernel.org/kvm/20240710074410.770409-1-suleiman@google.com/t/)

BUG=b:340428186
TEST=Manual testing with tracepoints to make sure host suspend is
    included in steal time, when using virtio-pvclock.

Change-Id: Id2cfd69071c875efa6489d26f0c0ec055c325315
Signed-off-by: Suleiman Souhlal <suleiman@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5740949
Reviewed-by: Vineeth Pillai <vineethrp@google.com>