CHROMIUM: arm64: dts: mt8186: Modify sbs battery reg to 0x0f
During the process of battery charging and discharging,
the display power will jump. This is because when reading
the power, the display power is calculated in watts due to
unsynchronized reading and writing. Use virtual battery 0xf
to communicate with EC can solve this.

BUG=b:354050945
TEST=check dispaly battery
UPSTREAM-TASK=b:354073363

Change-Id: I0e31cb3e2d0f11abd1352ba1b5edcb5b6d0c796b
Signed-off-by: wangganxiang <wangganxiang@huaqin.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5724209
Reviewed-by: cong yang <yangcong5@huaqin.corp-partner.google.com>
Commit-Queue: Hsin-Yi Wang <hsinyi@chromium.org>
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>